[DMP-435] Deleted tooltip navigate button props
