+++ src/components/common/Tooltip/Tooltip.module.scss
@@ -20,9 +20,3 @@
   top: 10px;
   right: 10px;
 }
-
-.tooltip-button {
-  display: flex;
-  justify-content: center;
-  margin-top: 20px;
-}
